Remove unnecessary casts
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/trunk@325313 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/org/apache/jmeter/engine/LagDetector.java
@@ -47,7 +47,7 @@ public class LagDetector extends Thread
      */
     public float getAveLag()
     {
-        return (float) ((float) totalLag / (float) count);
+        return ((float) totalLag / (float) count);
     }
 
     /**
